move definitions out of __init__
