+++ pupil_src/shared_modules/surface_tracker/surface_tracker_online.py
@@ -17,8 +17,7 @@ import pyglui
 import pyglui.cygl.utils as pyglui_utils
 import gl_utils
 
-
-from surface_tracker import Heatmap_Mode
+from surface_tracker.gui import Heatmap_Mode
 from surface_tracker.surface_tracker import Surface_Tracker
 from surface_tracker.surface_online import Surface_Online
